% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_connect.R
\name{jira_connect}
\alias{jira_connect}
\title{Validate the connection to jira by logging in an performing a search}
\usage{
jira_connect(host, user, password)
}
\arguments{
\item{host}{}

\item{user}{}

\item{password}{}
}
\description{
Validate the connection to jira by logging in an performing a search
}
