% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_utils.R
\name{jira_get}
\alias{jira_get}
\title{Runs a rest GET call towards jira}
\usage{
jira_get(url = url, user = user, password = password, verbose = verbose)
}
\arguments{
\item{url}{}

\item{user}{}

\item{password}{}

\item{verbose}{}
}
\value{
a parsed rest document (in the shape of a list)
}
\description{
Runs a rest GET call towards jira
}
\examples{

res <- jira_get(url = url, user = user, password = password)
}
