% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_utils.R
\name{map_chr_hack}
\alias{map_chr_hack}
\title{to avoid issues with map_chr and NULL}
\usage{
map_chr_hack(.x, .f, ...)
}
\arguments{
\item{.x}{}

\item{.f}{}

\item{...}{}
}
\value{
a character vector
}
\description{
to avoid issues with map_chr and NULL
}
\examples{
A helper function from https://github.com/jennybc/analyze-github-stuff-with-r#readme
same arguments as map_chr

  issue <-  res$issues \%>\% {
     tibble(issuetype_id = map_chr_hack(., ~ .x[["fields"]][["issuetype"]][["id"]]))} 
}
